Refactor object base even more.
